% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrNA.R
\name{attrNA}
\alias{attrNA}
\title{Atribuição de Preços Faltantes}
\usage{
attrNA(x)
}
\arguments{
\item{x}{Vetor organizado temporalmente.}
}
\value{
Vetor, do mesmo comprimento de \code{x}, com os valores faltantes
atribuídos, se possível.
}
\description{
Em uma série temporal \code{x}, atribui o valor de \code{x[i-1]} a \code{x[i]}, caso
\code{is.na(x[i])} seja \code{TRUE} e \code{is.na(x[i-1])} seja \code{FALSE}.
}
\examples{

x <- c(NA,NA,1,3,NA,7,-5,NA,NA,13)
print(attrNA(x))
}
