% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expShortfall.R
\name{expShortfall}
\alias{expShortfall}
\title{Expected Shortfall}
\usage{
expShortfall(rend, alpha = 0.05)
}
\arguments{
\item{rend}{Vetor de rendimentos.}

\item{alpha}{Quantil do VaR (default = 0.05).}
}
\value{
Déficit Esperado
}
\description{
Calcula o Déficit Esperado dado um vetor de rendimentos e um quantil. Enquanto o Value at
Risk (VaR) representa o retorno equivalente a um quantil \code{alpha}, o ES é a área sob a
curva de densidade de probabilidade dos retornos entre os quantis 0 e \code{alpha}.
}
\details{
O VaR apresenta o melhor retorno entre os piores cenários possíveis e o ES, o valor esperado
(soma ponderada pelas probabilidades de ocorrência) de todos os piores cenários possíveis.
}
\examples{

require(pracma)
require(magrittr)
require(DescTools)

set.seed(1)

y <- rnorm(100,0.08,0.30)

alfa <- 0.1

print(expShortfall(y,alfa))
}
\seealso{
\code{\link{VaRfunc}}
}
