% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finSummary.R
\name{finSummary}
\alias{finSummary}
\title{Sumarização das Métricas de Desempenho}
\usage{
finSummary(vec_rend, s, rf = 0)
}
\arguments{
\item{vec_rend}{Vetor com os rendimentos, organizados temporalmente.}

\item{s}{Lag, em dias corridos, sobre o qual foram calculados os rendimentos.}

\item{rf}{Taxa livre de risco relativa ao período dos rendimentos. Se os retornos foram
calculados com um lag de 120 dias, a rf deverá ser relativa a um período de 120 dias.}
}
\value{
Data frame contendo:
\enumerate{
\item Effective Annual Rate (EAR);
\item Desvio-padrão anualizado dos retornos;
\item Sharpe Ratio anualizado;
\item Drawdown em uma base de 30 dias. Se \code{s} é de 180 dias, por exemplo, o drawdown será
dividido por 6. O valor é correspondente ao quantil 5\% do vetor de rendimentos
\code{vec_rend}.
\item Taxa de períodos com retorno positivo.
}
}
\description{
Dado um vetor de retornos (principalmente de estratégias), exibe métricas de desempenho
anualizadas.
}
\examples{

set.seed(1)

rend <- rnorm(300,0.08,0.15)
print(finSummary(rend,180,0.03))
}
