% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMatRend.R
\name{genMatRend}
\alias{genMatRend}
\title{Matriz de Rendimentos Sincronizados}
\usage{
genMatRend(prices_list, lag_DC)
}
\arguments{
\item{prices_list}{\code{list} com data.frames no formato "data-preço"}

\item{lag_DC}{Tempo, em dias corridos, da avaliação dos retornos. O retorno será o de
composição contínua (equivalente a \code{TTR::ROC}):

\deqn{rcc = ln(P[D+lag_DC]/P[D+0])}}
}
\value{
Matriz com os retornos associados a cada pregão disponível
}
\description{
Dada uma lista cujos elementos são data.frames com duas colunas (a primeira
contendo as datas dos pregões e a segunda, os preços correspondentes) para cada ativo, é
gerada uma matriz de retornos (variações de preço) em que cada linha corresponde a uma data
de pregão e cada coluna, a um ativo.
}
\examples{

require(quantmod)
require(magrittr)

precos1 <- getSymbols("petr4.sa",
                       from="2020-01-01",to="2020-12-31",
                       auto.assign=FALSE)
precos2 <- getSymbols("vale3.sa",
                       from="2020-01-01",to="2020-12-31",
                       auto.assign=FALSE)

datas1 <- index(precos1)
datas2 <- index(precos2)

close1 <- precos1[,4]
close2 <- precos2[,4]

lista_precos <- list()
lista_precos[["PETR4"]] <- data.frame(data = datas1,fechamento = close1)
lista_precos[["VALE3"]] <- data.frame(data = datas2,fechamento = close2)

# 120 dias corridos
mat_rend <- genMatRend(lista_precos,120)
}
