% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markowitz.R
\name{markowitz}
\alias{markowitz}
\title{Otimização de Carteiras via Markowitz}
\usage{
markowitz(rend, param)
}
\arguments{
\item{rend}{Matriz com os rendimentos do mercado e de cada componente da carteira, nomeada.
O vetor com rendimentos do mercado não entra no cálculo da otimização, mas é necessário para
manter a padronização de cabeçalho entre as funções de otimização de carteira presentes na
biblioteca.}

\item{param}{\code{list} com os seguintes campos:

1: nº da coluna com rendimentos do mercado}
}
\value{
Vetor com as proporções de cada ativo na carteira ótima. A proporção do ativo
correspondente ao mercado será 0.
}
\description{
Encontra as proporções ótimas dos ativos na composição de uma carteira de investimentos,
usando as médias históricas de retorno e a covariância entre os retornos dos ativos.
}
\examples{

require(magrittr)

set.seed(1)

# vetor com rendimento do mercado
mkt <- rnorm(100,0.05,0.13)

# otimização com 4 ativos
stk1 <- rnorm(100,0.08,0.2)
stk2 <- rnorm(100,-0.03,0.05)
stk3 <- rnorm(100,0.13,0.45)
stk4 <- rnorm(100,0.02,0.1)

# matriz de rendimentos
mat_rend <- cbind(mkt,stk1,stk2,stk3,stk4)

# otimização: mercado na coluna 1
wopt <- markowitz(mat_rend,list(1))

# valores em \%
print(round(100*wopt,3))
}
