% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimSel.R
\name{optimSel}
\alias{optimSel}
\title{Seleção de Ativos e Otimização de Carteira}
\usage{
optimSel(rend, num_acoes, fun.otim, fun.sel, param.otim, param.sel)
}
\arguments{
\item{rend}{Matriz em que cada coluna representa as rentabilidades de um ativo individual.
As colunas devem ser nomeadas.}

\item{num_acoes}{Número de ativos que devem constar na carteira.}

\item{fun.otim}{Função de otimização de carteira.}

\item{fun.sel}{Função de seleção de ativos.}

\item{param.otim}{Parâmetros da função de otimização.}

\item{param.sel}{Parâmetros da função de seleção.}
}
\value{
\code{list} com os seguintes campos:
\enumerate{
\item tickers: nomes dos ativos selecionados;
\item w: porcentagem de cada ativo selecionado na carteira ótima.
}
}
\description{
Dada uma função de otimização de carteira e uma função de seleção de ativos, retorna a
composição da carteira ótima em uma matriz de rendimentos. Simplifica a aplicação de
metodologias de formação de portfólios.
}
\details{
A função de otimização deve ter 2 argumentos: a matriz de rendimentos e uma lista de
parâmetros (veja as funções \code{markowitz} e \code{sfm}).

A função de seleção de ativos deve ter 3 argumentos: a matriz de rendimentos, o número de
ações requeridas na carteira e uma lista de parâmetros (veja as funções \code{selBySR} e
\code{selByDelta}).
}
\examples{

require(magrittr)

# matriz de rendimentos
set.seed(1)

mi <- runif(11,-0.2,0.2)
sig <- runif(11,0.3,0.9)

mat_rend <- NULL
# 10 ativos + Mercado
for (i in 1:11){
   mat_rend <- cbind(mat_rend,
                     rnorm(300,mi,sig))
}
colnames(mat_rend) <- paste0("acao",1:11)

# aplicação
result <- optimSel(mat_rend,5,
                   markowitz,selBySR,
                   list(1),list(1,0.05))
print(result)
}
