% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selByDelta.R
\name{selByDelta}
\alias{selByDelta}
\title{Seleção de Ativos Pelo Delta}
\usage{
selByDelta(rend, num_acoes, param)
}
\arguments{
\item{rend}{Matriz em que cada coluna representa as rentabilidades de um ativo individual.
As colunas devem ser nomeadas.}

\item{num_acoes}{Quantas ações devem constar na carteira?}

\item{param}{\code{list} contendo os seguintes campos:
\enumerate{
\item nº da coluna com rendimentos do mercado.
}}
}
\value{
Vetor com os \code{num_acoes} tickers (nomes dos ativos) com maior valor de Delta,
além do ticker equivalente ao Mercado.
}
\description{
Delta é o desempenho acima do previsto pela regressão linear dos retornos do ativo com os
retornos do Mercado. A diferença entre os retornos real e previsto é tomada no último preço
disponível.
}
\examples{

require(magrittr)

set.seed(1)

mi <- runif(11,-0.2,0.2)
sig <- runif(11,0.3,0.9)

mat_rend <- NULL
# 10 ativos + Mercado
for (i in 1:11){
   mat_rend <- cbind(mat_rend,
                     rnorm(300,mi,sig))
}
colnames(mat_rend) <- paste0("acao",1:11)

print(selByDelta(mat_rend,6,list(1)))
}
