% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selBySR.R
\name{selBySR}
\alias{selBySR}
\title{Seleção de Ativos Pela Relação Risco-Retorno}
\usage{
selBySR(rend, num_acoes, param)
}
\arguments{
\item{rend}{Matriz em que cada coluna representa as rentabilidades de um ativo individual.
As colunas devem ser nomeadas.}

\item{num_acoes}{Quantas ações devem constar na carteira?}

\item{param}{\code{list} contendo os seguintes campos:
\enumerate{
\item nº da coluna com rendimentos do mercado;
\item taxa livre de risco para o lag sobre o qual foram calculados os retorno.
}}
}
\value{
Vetor com os \code{num_acoes} tickers (nomes dos ativos) com maior razão
RETORNO/RISCO, além do ticker equivalente ao Mercado.
}
\description{
Ordena os ativos cujos retornos são passados para a função por ordem decrescente da razão
RETORNO/RISCO.
}
\examples{

require(magrittr)

set.seed(1)

mi <- runif(11,-0.2,0.2)
sig <- runif(11,0.3,0.9)

mat_rend <- NULL
# 10 ativos + Mercado
for (i in 1:11){
   mat_rend <- cbind(mat_rend,
                     rnorm(300,mi,sig))
}
colnames(mat_rend) <- paste0("acao",1:11)

print(selBySR(mat_rend,6,list(1,0.05)))
}
