% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm.R
\name{sfm}
\alias{sfm}
\title{Otimização de Carteiras via Modelo de Fator Único}
\usage{
sfm(rend, param)
}
\arguments{
\item{rend}{Matriz com os rendimentos do mercado e de cada componente da carteira, nomeada.}

\item{param}{\code{list} com os seguintes campos:

1: nº da coluna com rendimentos do mercado

2: máxima alavancagem. Se |alavancagem| <= 1, não há operações short}
}
\value{
Vetor com as proporções de cada ativo na carteira ótima. Proporções negativas
indicam posições vendidas.
}
\description{
Encontra as proporções ótimas dos ativos na composição de uma carteira de investimentos,
usando os coeficientes "beta" dos ativos e seus riscos específico e sistemático. Considera a
possibilidade de posições compradas e vendidas e alavancagem.
}
\examples{

require(magrittr)

set.seed(1)

# vetor com rendimento do mercado
mkt <- rnorm(100,0.05,0.13)

# otimização com 4 ativos
stk1 <- rnorm(100,0.08,0.2)
stk2 <- rnorm(100,-0.03,0.05)
stk3 <- rnorm(100,0.13,0.45)
stk4 <- rnorm(100,0.02,0.1)

# matriz de rendimentos
mat_rend <- cbind(mkt,stk1,stk2,stk3,stk4)

# otimização: mercado na coluna 1
wopt <- sfm(mat_rend,list(1,2))

# valores em \%
print(round(100*wopt,3))
}
