% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{sharpe}
\alias{sharpe}
\title{Sharpe Ratio}
\usage{
sharpe(x, rf = 0)
}
\arguments{
\item{x}{Vetor com os rendimentos.}

\item{rf}{Taxa livre de risco relativa ao período dos rendimentos. Se os retornos foram
calculados com um lag de 120 dias, a rf deverá ser relativa a um período de 120 dias.}
}
\value{
SR dos retornos presentes em \code{x}.
}
\description{
Calcula o índice de Sharpe de uma série de retornos.
}
\examples{

set.seed(1)

rend <- rnorm(300,0.08,0.15)
print(sharpe(rend,0.03))
}
