% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStrat.R
\name{testStrat}
\alias{testStrat}
\title{Testes t (médias) e F (variâncias) Entre Retornos}
\usage{
testStrat(rend)
}
\arguments{
\item{rend}{Matriz em que cada coluna representa as rentabilidades de uma estratégia ou
ativo individual. As colunas devem ser nomeadas.}
}
\value{
Lista contendo os seguintes campos:
\enumerate{
\item relações entre os retornos;
\item relações entre as variâncias.
}

Os elementos dessa matriz são os termos "greater", "less" e "equal", indicando se a primeira
medida testada é maior que, menor que ou igual à, em termos de média e variância, segunda.

Sendo 'n' o número de colunas de \code{rend}, os dois elemento dessa lista são matrizes
triangulares (n-1) x (n-1). A ordem de leitura é, para o elemento \eqn{[i,j]} da matriz de
saída M: "retorno/volatilidade da estratégia i é maior que/menor que/igual a (de acordo com
\eqn{M[i,j]}) retorno/volatilidade da estratégia j".
}
\description{
Avalia a relação (maior, menor ou igual) entre pares de rentabilidades, em termos de média
dos retornos (associada à esperança de retorno futura) e variância dos retornos (associada
ao risco da estratégia).
}
\details{
Os teste são realizados com nível de confiança de 95\%.
}
\examples{

set.seed(1)

mat_rend <- cbind(estrat1 = rnorm(300,0.08,0.2),
                  estrat2 = rnorm(300,0.08,0.4),
                  estrat3 = rnorm(300,0.04,0.2))

print(testStrat(mat_rend))
}
