% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaRfunc.R
\name{VaRfunc}
\alias{VaRfunc}
\title{Value at Risk}
\usage{
VaRfunc(rend, alpha = 0.05)
}
\arguments{
\item{rend}{Vetor de rendimentos.}

\item{alpha}{Quantil do VaR (default = 0.05).}
}
\value{
Valor em Risco.
}
\description{
Calcula o Valor em Risco dado um vetor de rendimentos e um quantil. O VaR é
equivalente à melhor taxa de retorno dentre os \code{alpha} piores cenários futuros. A
expectativa é que \code{1-alpha} retornos futuros superem o VaR.
}
\details{
O valor do quantil é equivalente a 1 - Nível de Confiança.
}
\examples{

require(pracma)

set.seed(1)

y <- rnorm(100,0.08,0.30)

alfa <- 0.1

print(VaRfunc(y,alfa))
}
